test(cli): convert flag validation test to a unit test
